# swagger.render.json
{
  "swagger": "2.0",
  "info": {
    "title": "The Ultimate Music Database",
    "description": "The final project for CSE341",
    "version": "1.0.0"
  },
  "host": "cse341-tumd.onrender.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
    "/profile": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/": {
      "get": {
        "tags": [
          "Home"
        ],
        "description": "This is the home page. It does not do much; atleast not so far.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/front/home": {
      "get": {
        "tags": [
          "Front"
        ],
        "description": "Home page of the front endpoint",
        "parameters": [],
        "responses": {}
      }
    },
    "/songs": {
      "get": {
        "tags": [
          "Song"
        ],
        "description": "Get list of all songs",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Song"
        ],
        "description": "Insert new song to the collection",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "required": [
                "title",
                "release_year",
                "summary"
              ],
              "properties": {
                "title": {
                  "type": "string",
                  "example": "Some title..."
                },
                "release_year": {
                  "type": "string",
                  "example": "Some release year..."
                },
                "rating": {
                  "type": "number",
                  "example": "Some rating..."
                },
                "summary": {
                  "type": "string",
                  "example": "Some summary..."
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/songs/{song_id}": {
      "get": {
        "tags": [
          "Song"
        ],
        "description": "Get a specific song by song_id",
        "parameters": [
          {
            "name": "song_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The ID of the song to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Song"
        ],
        "description": "Updates a song specified by song_id",
        "parameters": [
          {
            "name": "song_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The ID of the song to update"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "required": [
                "title",
                "release_year",
                "summary"
              ],
              "properties": {
                "title": {
                  "type": "string",
                  "example": "Some title..."
                },
                "release_year": {
                  "type": "string",
                  "example": "Some release year..."
                },
                "rating": {
                  "type": "number",
                  "example": "Some rating..."
                },
                "summary": {
                  "type": "string",
                  "example": "Some summary..."
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Song"
        ],
        "description": "Deletes a song specified by song_id",
        "parameters": [
          {
            "name": "song_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The ID of the song to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/artists": {
      "get": {
        "tags": [
          "Artist"
        ],
        "description": "Get list of all artists",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Artist"
        ],
        "description": "Insert new artist to the collection",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "required": [
                "first_name",
                "last_name",
                "gender"
              ],
              "properties": {
                "first_name": {
                  "type": "string",
                  "example": "Some title..."
                },
                "last_name": {
                  "type": "string",
                  "example": "Some release year..."
                },
                "gender": {
                  "type": "string",
                  "example": "Some rating..."
                },
                "date_of_birth": {
                  "type": "date",
                  "example": "Some summary..."
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/artists/{artist_id}": {
      "get": {
        "tags": [
          "Artist"
        ],
        "description": "Get a specific artist by artist_id",
        "parameters": [
          {
            "name": "artist_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The ID of the artist to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Artist"
        ],
        "description": "Updates a artist specified by artist_id",
        "parameters": [
          {
            "name": "artist_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The ID of the artist to update"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "required": [
                "first_name",
                "last_name",
                "gender"
              ],
              "properties": {
                "first_name": {
                  "type": "string",
                  "example": "Some title..."
                },
                "last_name": {
                  "type": "string",
                  "example": "Some release year..."
                },
                "gender": {
                  "type": "string",
                  "example": "Some rating..."
                },
                "date_of_birth": {
                  "type": "date",
                  "example": "Some summary..."
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Artist"
        ],
        "description": "Deletes a artist specified by artist_id",
        "parameters": [
          {
            "name": "artist_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The ID of the artist to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/reviews": {
      "get": {
        "tags": [
          "Review"
        ],
        "description": "Get list of all reviews",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Review"
        ],
        "description": "Insert new review to the collection",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "required": [
                "body",
                "rating",
                "user"
              ],
              "properties": {
                "body": {
                  "type": "string",
                  "example": "Some title..."
                },
                "rating": {
                  "type": "number",
                  "example": 4.5
                },
                "user": {
                  "type": "string"
                },
                "song": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/reviews/{review_id}": {
      "get": {
        "tags": [
          "Review"
        ],
        "description": "Get a specific review by review_id",
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The ID of the review to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Review"
        ],
        "description": "Updates a review specified by review_id",
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The ID of the review to retrieve"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "required": [
                "body",
                "rating",
                "user"
              ],
              "properties": {
                "body": {
                  "type": "string",
                  "example": "Some title..."
                },
                "rating": {
                  "type": "number",
                  "example": 4.5
                },
                "user": {
                  "type": "string"
                },
                "song": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "delete": {
        "tags": [
          "Review"
        ],
        "description": "Deletes a review specified by review_id",
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The ID of the review to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/genres": {
      "get": {
        "tags": [
          "Genre"
        ],
        "description": "Get list of all genres",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Genre"
        ],
        "description": "Insert new genre to the collection",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "required": [
                "name"
              ],
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Some title..."
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/genres/{genre_id}": {
      "get": {
        "tags": [
          "Genre"
        ],
        "description": "Get a specific genre by genre_id",
        "parameters": [
          {
            "name": "genre_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The ID of the genre to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Genre"
        ],
        "description": "Updates a genre specified by genre_id",
        "parameters": [
          {
            "name": "genre_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The ID of the genre to update"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "required": [
                "name"
              ],
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Some title..."
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Genre"
        ],
        "description": "Deletes a genre specified by genre_id",
        "parameters": [
          {
            "name": "genre_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The ID of the genre to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Get list of all users. Only displays fields that can be considered safe for public. Fields like email address, etc are skipped",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users/{user_id}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Get a specific user by user_id",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The ID of the user to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  }
}